20220226_[1.0.0]_RealseVer

+++ src/main/java/com/thaithong/datn/model/UserAssignmentResponseModel.java
@@ -31,7 +31,7 @@ public class UserAssignmentResponseModel {
      */
     private Long responseId;
 
-    private Double rate;
+    private Integer rate;
 
     private Boolean isCompleted;
